Multi node

+++ scripts/install_wmlce.sh
@@ -17,7 +17,23 @@
 source /tmp/scripts/env.sh
 
 #Install Miniconda
-wget https://repo.anaconda.com/miniconda/Miniconda3-latest-Linux-ppc64le.sh
+#Sometimes wget fails to resolve repo.anaconda.com, that's why, adding a retry loop
+retry=0
+maxRetries=5
+retryInterval=5
+until [ ${retry} -ge ${maxRetries} ]
+do
+	wget https://repo.anaconda.com/miniconda/Miniconda3-latest-Linux-ppc64le.sh && break
+	retry=$[${retry}+1]
+	echo "Retrying [${retry}/${maxRetries}] in ${retryInterval}(s) "
+	sleep ${retryInterval}
+done
+
+if [ ${retry} -ge ${maxRetries} ]; then
+  echo "Failed after ${maxRetries} attempts!"
+  exit 1
+fi
+
 bash Miniconda3-latest-Linux-ppc64le.sh -b -f
 rm Miniconda3-latest-Linux-ppc64le.sh
 
@@ -25,7 +41,7 @@ rm Miniconda3-latest-Linux-ppc64le.sh
 eval "$($HOME/miniconda3/bin/conda shell.bash hook)"
 
 #Setup conda environment
-conda config --prepend channels https://wml-ce-proxy-cos.s3.direct.us-east.cloud-object-storage.appdomain.cloud/ 
+conda config --prepend channels https://wml-ce-proxy-cos.s3.direct.us-east.cloud-object-storage.appdomain.cloud/
 conda create --name wmlce_env_${WMLCE_VERSION}  python=${PYTHON_VERSION} -y
 conda activate wmlce_env_${WMLCE_VERSION}
